RIrtWrappers.jl

This package wraps some R packages for fitting IRT models and running CATs.

Fitting IRT models

The KernSmoothIRT and Mirt modules wrap R libraries for fitting IRT models.

The models are returned as item banks as in FitttedItemBanks.jl. The inputs are response matrices, in the same format as provided by ItemResponseDatasets.jl. In particular a DataFrames.DataFrame with questions as columns and respondents as rows, with outcomes 0-based integer coded.

For example:

4×4 DataFrame
   Row │ Q1     Q2     Q3     Q4
       │ Int64  Int64  Int64  Int64
───────┼───────────────────────────
     1 │     1      1      0      0
     2 │     1      0      1      0
     3 │     1      1      1      1
     4 │     1      1      0      1

Running CATs

The MirtCat module wraps an R library for simulating and administering CATs.

You need to use the following method to get ahold of the module.

RIrtWrappers.require_mirtcat — Function

require_mirtcat() -> Module

Returns the MirtCat extension module. Requires the ComputerAdaptiveTesting module in your environment.

The main objects from these always implement the ComputerAdaptiveTesting.Stateful interface.

Index

RIrtWrappers.KernSmoothIRT

RIrtWrappers.Mirt

MirtCat.MirtCatDesign

MirtCat.StatefulMirtCat

MirtCat.StatefulMirtCatNoRollbacks

MirtCat.StatefulMirtCatWithRollbacks

MirtCat.make_mirtcat

RIrtWrappers.KernSmoothIRT.fit_ks_dichotomous

RIrtWrappers.Mirt.fit_2pl

RIrtWrappers.Mirt.fit_3pl

RIrtWrappers.Mirt.fit_4pl

RIrtWrappers.Mirt.fit_gpcm

RIrtWrappers.Mirt.fit_mirt_2pl

RIrtWrappers.Mirt.fit_monopoly

RIrtWrappers.Mirt.fit_spline

RIrtWrappers.require_mirtcat